Rename compute shaders to .glsl ending

+++ shaders/julia_gradient.comp.glsl
@@ -12,20 +12,20 @@ layout(binding = 0, rgba8) writeonly uniform image2D colorTex;
 
 int MAX_ITERATIONS = 100;
 
-double julia(dvec2 z, dvec2 c) {
+float julia(vec2 z, vec2 c) {
 	for (int i = 0; i < MAX_ITERATIONS; i++) {
-		dvec2 r = z * z;
+		vec2 r = z * z;
 		if (r.x + r.y > 4.0) return i;
-		z = dvec2(r.x - r.y + c.x, 2.0 * z.x * z.y + c.y);
+		z = vec2(r.x - r.y + c.x, 2.0 * z.x * z.y + c.y);
 	}
 	return MAX_ITERATIONS;
 }
 
-vec3 backgroundGradient(double colour, ivec2 pixelCoords) {
+vec3 backgroundGradient(float colour, ivec2 pixelCoords) {
 	vec2 pt = pixelCoords;
-	dvec3 color1 = dvec3(1.0, 0.55, 0.0) * colour;
-	dvec3 color2 = dvec3(0.226, 0.0, 0.615) * colour;
-	double mixValue = distance(pt, vec2(0,1));
+	vec3 color1 = vec3(1.0, 0.55, 0.0) * colour;
+	vec3 color2 = vec3(0.226, 0.0, 0.615) * colour;
+	float mixValue = distance(pt, vec2(0,1));
 	return vec3(mix(color1, color2, mixValue));
 }
 
@@ -38,11 +38,11 @@ void main() {
 
 	int m = min(iResolution.x, iResolution.y);
 	ivec2 mdiff = (iResolution - m) / 2;
-	dvec2 coords = dvec2(translateCoordinates(pixelCoords.x, m + mdiff.x, mdiff.x),
-						 translateCoordinates(pixelCoords.y, m + mdiff.y, mdiff.y));
+	vec2 coords = vec2(translateCoordinates(pixelCoords.x, m + mdiff.x, mdiff.x),
+	translateCoordinates(pixelCoords.y, m + mdiff.y, mdiff.y));
 
 	// https://en.wikipedia.org/wiki/Julia_set for more examples to put instead of dvec2(-0.5, 0.156)
-	double fractal = julia(coords.xy, dvec2(-0.8, 0.156)) / MAX_ITERATIONS;
+	float fractal = julia(coords.xy, vec2(-0.8, 0.156)) / MAX_ITERATIONS;
 	vec3 colourOut = backgroundGradient(fractal, pixelCoords / iResolution.xy);
 
 	imageStore(colorTex, pixelCoords, vec4(colourOut, 1.0));
